+++ Cmdlets/Private/Resolve-TeamViewerUserGroupMemberId.ps1
@@ -7,16 +7,16 @@ function Resolve-TeamViewerUserGroupMemberMemberId {
 
     process {
         if ($UserGroupMember.PSObject.TypeNames -contains 'TeamViewerPS.UserGroupMember') {
-            $UserGroupMember.AccountId
+            Write-Output $UserGroupMember.AccountId
         }
         elseif ($UserGroupMember -match 'u[0-9]+') {
-            $UserGroupMember
+            Write-Output $UserGroupMember
         }
         elseif ($UserGroupMember -is [string]) {
-            [int]$UserGroupMember
+            Write-Output ([int]$UserGroupMember)
         }
         elseif ($UserGroupMember -is [int]) {
-            $UserGroupMember
+            Write-Output $UserGroupMember
         }
         else {
             throw "Invalid user group identifier '$UserGroupMember'. Must be either a [TeamViewerPS.UserGroupMember],[TeamViewerPS.User] or [int] ."