+++ Cmdlets/Public/Add-TeamViewerManagedDevice.ps1
@@ -1,5 +1,6 @@
 function Add-TeamViewerManagedDevice {
     [CmdletBinding(SupportsShouldProcess = $true)]
+
     param(
         [Parameter(Mandatory = $true)]
         [securestring]
@@ -7,13 +8,13 @@ function Add-TeamViewerManagedDevice {
 
         [Parameter(Mandatory = $true)]
         [ValidateScript( { $_ | Resolve-TeamViewerManagedDeviceId } )]
-        [Alias("DeviceId")]
+        [Alias('DeviceId')]
         [object]
         $Device,
 
         [Parameter(Mandatory = $true)]
         [ValidateScript( { $_ | Resolve-TeamViewerManagedGroupId } )]
-        [Alias("GroupId")]
+        [Alias('GroupId')]
         [object]
         $Group
     )
@@ -23,17 +24,16 @@ function Add-TeamViewerManagedDevice {
     $resourceUri = "$(Get-TeamViewerApiUri)/managed/groups/$groupId/devices"
 
     $body = @{
-        id = $deviceId
+        id = $deviceId.ToString()
     }
 
-    if ($PSCmdlet.ShouldProcess($deviceId, "Add device to managed group")) {
+    if ($PSCmdlet.ShouldProcess($deviceId, 'Add device to managed group')) {
         Invoke-TeamViewerRestMethod `
             -ApiToken $ApiToken `
             -Uri $resourceUri `
             -Method Post `
-            -ContentType "application/json; charset=utf-8" `
+            -ContentType 'application/json; charset=utf-8' `
             -Body ([System.Text.Encoding]::UTF8.GetBytes(($body | ConvertTo-Json))) `
-            -WriteErrorTo $PSCmdlet | `
-            Out-Null
+            -WriteErrorTo $PSCmdlet | Out-Null
     }
 }